+++ src/components/game/Sketch.tsx
@@ -2,12 +2,13 @@ import p5 from 'p5'
 import p5Types from 'p5'
 import React, { useState } from 'react'
 import Sketch from 'react-p5'
-import Ending, { RestartButton } from './Ending'
 import 'p5/lib/addons/p5.sound'
 import { Evaluation, interpret, WRENatives, Id } from 'wollok-ts'
 import { GameProject, DEFAULT_GAME_ASSETS_DIR } from './gameProject'
 import { Board, boardToLayers } from './utils'
 import { flushEvents, boardGround, cellSize, width, height, currentSoundStates, SoundState, io, nextBoard, gameStop } from './GameStates'
+import { Button } from '@material-ui/core'
+import ReplayIcon from '@material-ui/icons/Replay'
 
 const defaultImgs = [
   'ground.png',
@@ -224,14 +225,17 @@ const SketchComponent = ({ game, evaluation: e }: SketchProps) => {
     queueGameEvent(anyKeyPressedId)
     return false
   }
-
-  return stop
-    ? <Ending restartProps={{ restart }} />
-    : <div>
-      <Sketch setup={setup as any} draw={draw as any} keyPressed={keyPressed as any} />
-      <RestartButton restart={restart} />
-    </div>
-
+  return <div>
+    {stop ?
+      <h1>Se terminó el juego</h1>
+      : <Sketch setup={setup as any} draw={draw as any} keyPressed={keyPressed as any} />}
+    <RestartButton restart={restart} />
+  </div>
 }
 
 export default SketchComponent
+
+type RestartProps = { restart: () => void }
+export function RestartButton(props: RestartProps) {
+  return <Button onClick={event => { event.preventDefault(); props.restart() }} variant="contained" color="primary" startIcon={<ReplayIcon />}>Reiniciar el juego</Button>
+}
